Branch: feature/3
task 1 Product name cannot be less than 3 characters and more than 32. ok
task 2 The discount should not be more than 100%. ok
task 3 The discount cannot be less than 0. The discount may be a fractional number (10.5%). ok (discount type is integer)
task 4 The price of the product must be greater than 0. ok
task 7 If the price of the product is less than $ 20, then a discount cannot be made. ok
task 6 The name of the product must be unique. failed